Handle checkboxes and add button
