+++ src/components/Voyage/FormVoyageVessels.jsx
@@ -7,42 +7,15 @@ import FormVessel from '@components/Vessel/FormVessel';
 import VesselTable from '@components/Vessel/VesselTable';
 
 const FormVessels = ({
-  handleSubmit, handleChange, errors, formData,
+  handleSubmit, handleChange, handleCheckboxes, handleAddButton, errors, formData,
 }) => {
   const [vesselData, setVesselData] = useState();
-  const [selectedVessel, setSelectedVessel] = useState();
 
   const storeVesselData = () => {
     getData(VESSELS_URL)
       .then((resp) => setVesselData(resp.items));
   };
 
-  // Handle checkboxes being checked/unchecked
-  const handleCheckboxes = (e) => {
-    if ((e.target).checked) {
-      // Get this vessel data
-      getData(`${VESSELS_URL}/${e.target.id}`)
-        // Overwrite checkedVesselData with this vesselData
-        .then((resp) => setSelectedVessel(resp));
-    }
-    // Uncheck every other option
-    const vesselCheckboxes = document.querySelectorAll('input[name=vessel]');
-    Array.from(vesselCheckboxes).map((vessel) => {
-      if (e.target.id !== vessel.id && vessel.checked) {
-        vessel.checked = false;
-      }
-    });
-  };
-
-
-  // Handle 'add to report' to populate vessel form below
-  const handleAdd = () => {
-    // Overwrite formData with vesselData
-    console.log(selectedVessel)
-  };
-
-  // Display vessel form
-  // Save and continue saves to this voyage only (does not update vessel)
 
   useEffect(() => {
     storeVesselData();
@@ -67,7 +40,7 @@ const FormVessels = ({
         type="button"
         className="govuk-button"
         data-module="govuk-button"
-        onClick={(e) => handleAdd(e)}
+        onClick={(e) => handleAddButton(e)}
       >
         Add to report
       </button>
